update: tsファイル名前入れ替え
reading.tsにクイズの単語を決めるコードを、quiz.tsに読み取りクイズのコードを書いてしまっていたので、それぞれのファイル名を変更しました
