+++ src/pages/quiz/index.html
@@ -5,7 +5,7 @@
     <title>Quiz</title>
     <link rel="stylesheet" href="../styles/style.css" />
     <link rel="stylesheet" href="./styles/quiz.css" />
-    <script type="module" src="./scripts/quiz.ts" defer></script>
+    <script type="module" src="./scripts/reading.ts" defer></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
