+++ src/pages/quiz/scripts/quiz.ts
@@ -1,165 +1,91 @@
-import { startQuiz, type QuizData } from "./reading.js";
-import { type ShuwaData } from "../../../types/index.js";
 import data from "../../../../data/shuwa.json";
-import VideoPlayer from "../../../components/video/video-player.js";
-// shuwa.jsonのデータ構造を仮定（実際の構造に合わせて変更してください）
-// interface ShuwaData {
-//   id: number;
-//   name: string;
-//   youtube_url: string;
-// }
+import type { ShuwaData } from "../../../types/index.js";
 
-// --- DOM要素の取得 ---
-const videoContainer = document.getElementById(
-  "video-container",
-) as HTMLDivElement;
-const choiceButtons = [
-  document.getElementById("choice1") as HTMLInputElement,
-  document.getElementById("choice2") as HTMLInputElement,
-  document.getElementById("choice3") as HTMLInputElement,
-  document.getElementById("choice4") as HTMLInputElement,
-];
-const retireButton = document.getElementById(
-  "retireButton",
-) as HTMLInputElement;
-
-// --- モーダル関連のDOM要素 (HTMLに追加する必要があり) ---
-const modal = document.getElementById("result-modal") as HTMLDivElement;
-const modalMessage = document.getElementById(
-  "modal-message",
-) as HTMLParagraphElement;
-const nextButton = document.getElementById("next-button") as HTMLButtonElement;
-
-// --- クイズの状態管理 ---
-let currentQuestionIndex = 0; //現在何問目かを記録
-let quizData: QuizData | null = null; //問題と選択肢を保存
-let allShuwaData: ShuwaData[] = []; //手話の全データ保存
-const results: boolean[] = []; // クイズの正負を保存（正ならtrue、負ならfalse）
-const quizIds: string[] = [];
-
-// --- メイン処理 ---
-document.addEventListener("DOMContentLoaded", async () => {
-  // shuwa.json全体を最初に読み込んでおく
-  allShuwaData = data as ShuwaData[];
-  quizData = await startQuiz("some-mode"); // 'some-mode'は適切なモード名に
-  if (quizData) {
-    displayQuestion();
-  }
-});
-
-// --- 関数定義 ---
-
-// 問題と選択肢を画面に表示する関数
-function displayQuestion() {
-  if (!quizData || currentQuestionIndex >= quizData.quizWords.length) {
-    // 全問終了
-      showFinalResult();
-    return;
-  }
-
-  const questionId = quizData.quizWords[currentQuestionIndex];
-  const choices = quizData.choices[currentQuestionIndex];
-  const questionVideoUrl = findDataById(questionId)?.youtube_url;
-
-  // 問題動画を表示
-  if (questionVideoUrl) {
-    videoContainer.innerHTML = VideoPlayer(questionVideoUrl);
-  }
-
-  // 選択肢ボタンに単語とIDを割り当て
-  choiceButtons.forEach((button, index) => {
-    const choiceId = choices[index];
-    const choiceWord = findDataById(choiceId)?.name;
-    button.value = choiceWord || "エラー";
-    // data属性にIDを保存しておくのが便利
-    button.dataset.choiceId = choiceId.toString();
-  });
+// クイズのデータ構造を定義
+export interface QuizData {
+  quizWords: number[];
+  choices: number[][];
 }
 
-// 選択肢ボタンのクリックイベント
-choiceButtons.forEach((button) => {
-  button.addEventListener("click", (event) => {
-    const target = event.target as HTMLInputElement;
-    const selectedId = parseInt(target.dataset.choiceId || "0", 10);
-    checkAnswer(selectedId);
-  });
-});
+export async function startQuiz(_mode: string): Promise<QuizData | null> {
+  let quizWords: number[] = []; //問題格納用配列
+  const choices: number[][] = []; // 各問題の選択肢を格納する2次元配列
 
-// 答え合わせをする関数
-function checkAnswer(selectedId: number) {
-  if (!quizData) return;
+  console.log("reading.tsを読み込みました"); // テスト用
 
-  const correctId = quizData.quizWords[currentQuestionIndex];
-  const isCorrect = selectedId === correctId;
+  try {
+    // わんちゃんエラーが起きそうな所
+    //--jsonファイルの大きさ取得----
+    const jsonData: ShuwaData[] = data as ShuwaData[];
+    // 型安全な方法でJSONデータを取得
 
-  // 結果を保存
-  results.push(isCorrect);
-  localStorage.setItem("quizResults", JSON.stringify(results));
-  //ローカルストレージにクイズの正負を保存
+    const dataCount = jsonData.length;
+    //-----------
+    //--Fisher-Yatesアルゴリズムで問題ランダム出題--
 
-  // 問題idを保存
-  quizIds.push(correctId.toString());
-  localStorage.setItem("quizIds", JSON.stringify(quizIds));
-  // モーダルで結果表示
-  showResultModal(isCorrect, correctId, selectedId);
-}
+    const numberPool: number[] = [];
+    // 簡単に言うとくじ引きの箱を作成
+    for (let i = 0; i < dataCount; i++) {
+      numberPool.push(i + 1); // 1から始まるように調整
+      // jsonにかかれているidは1からスタートのため
+    }
 
-// 結果表示モーダルを表示
-function showResultModal(
-  isCorrect: boolean,
-  correctId: number,
-  selectedId: number,
-) {
-  modalMessage.textContent = isCorrect ? "正解！" : "不正解...";
+    // 配列をシャッフルする（フィッシャー–イェーツのシャッフル）
+    for (let i = numberPool.length - 1; i > 0; i--) {
+      // 0からiまでのランダムなインデックスを生成
+      const j = Math.floor(Math.random() * (i + 1));
+      // 要素を交換する
+      [numberPool[i], numberPool[j]] = [numberPool[j], numberPool[i]];
+    }
 
-  const correctVideoUrl = findDataById(correctId)?.youtube_url;
-  const selectedVideoUrl = findDataById(selectedId)?.youtube_url;
+    // シャッフルされた配列の先頭から7個を取得して、問題用の配列に格納
+    //    slice(0, 7) は、0番目から7個分の要素を新しい配列として取り出す
+    quizWords = numberPool.slice(0, 5); //データができ次第、slice(0,10)にする
 
-  if (!correctVideoUrl || !selectedVideoUrl) return;
+    // --- 選択肢の生成 ---
+    // 全問題のIDリストを作成 (1からdataCountまで)
+    const allIds = Array.from({ length: dataCount }, (_, i) => i + 1);
 
-  const correctAnswerVideo = document.getElementById(
-    "correct-answer-video",
-  ) as HTMLDivElement;
-  const yourAnswerVideo = document.getElementById(
-    "your-answer-video",
-  ) as HTMLDivElement;
+    for (const correctAnswerId of quizWords) {
+      // 1. 正解以外の選択肢候補をフィルタリング
+      const wrongChoiceCandidates = allIds.filter(
+        (id) => id !== correctAnswerId,
+      );
 
-  correctAnswerVideo.innerHTML = VideoPlayer(correctVideoUrl);
-  yourAnswerVideo.innerHTML = VideoPlayer(selectedVideoUrl);
+      // 2. 不正解の選択肢候補をシャッフル
+      for (let i = wrongChoiceCandidates.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [wrongChoiceCandidates[i], wrongChoiceCandidates[j]] = [
+          wrongChoiceCandidates[j],
+          wrongChoiceCandidates[i],
+        ];
+      }
 
-  modal.style.display = "flex";
-}
+      // 3. 不正解の選択肢を3つ選ぶ
+      const wrongChoices = wrongChoiceCandidates.slice(0, 3);
 
-// 「次へ」ボタンのクリックイベント
-nextButton.addEventListener("click", () => {
-  modal.style.display = "none";
-  currentQuestionIndex++;
-  displayQuestion();
-});
+      // 4. 正解と不正解を合わせて選択肢リストを作成
+      const currentChoices = [correctAnswerId, ...wrongChoices];
 
-// IDから手話データを検索するヘルパー関数
-function findDataById(id: number): ShuwaData | undefined {
-  // Try to find by ID first
-  const found = allShuwaData.find((item) => item.id === id);
-  if (found) return found;
-  // Fallback: if id-1 is a valid index, return that item
-  if (id > 0 && id <= allShuwaData.length) {
-    return allShuwaData[id - 1];
+      // 5. 選択肢の順番をシャッフル
+      for (let i = currentChoices.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [currentChoices[i], currentChoices[j]] = [
+          currentChoices[j],
+          currentChoices[i],
+        ];
+      }
+      choices.push(currentChoices);
+    }
+
+    console.log(`生成された問題（重複なしランダム7件）:`, quizWords);
+    console.log(`生成された選択肢:`, choices);
+
+    // 生成した問題と選択肢のセットを返す
+    return { quizWords, choices };
+  } catch (error) {
+    // エラーが起きたときの処理
+    console.error("jsonファイルが読み込めませんでした", error);
+    return null;
   }
-  return undefined;
 }
-
-// 最終結果の表示（例）
-function showFinalResult() {
-  console.log("クイズ終了");
-  // タイトル画面などに戻る処理
-  //
-  window.location.href = "../result/";
-}
-
-// 終了ボタン
-retireButton.addEventListener("click", () => {
-  if (confirm("クイズを中断してタイトルに戻りますか？")) {
-    window.location.href = "../"; // タイトル画面のパス
-  }
-});
